Coach Schedule + Hide admins actions
I added the page for coaches which shows them which scheduled events they are enrolled in, and which members are enrolled in them.

I also added some additional isAdmin check for all details page as they previously gave a link to the edit page.

I also added the nav items back as I think they are more useful and show be shown to everyone....